Check if notify is called on same thread as Lua runs
Not just main thread. Also this is portable. We will use pthread-win32 as well so…

The reason for this code is libspotify sometimes asks for notify_main_thread to occur ON the main thread :P Our startup sequence means Spotify initializes before we call listen() on the socket, which means our blocking tellmate() holds up startup for > 5 seconds waiting for nothing.